middleware/requirements.txt
paho-mqtt>=1.6,<2
pyyaml>=6.0
requests>=2.28
watchdog>=3.0
websocket-client>=1.6
